+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Loading indicator in search field during name search [PR #234](https://github.com/marmot-protocol/whitenoise/pull/234)
 - Pass metadata to start chat screen for instant display [PR #234](https://github.com/marmot-protocol/whitenoise/pull/234)
 - Invite callout [PR #230](https://github.com/marmot-protocol/whitenoise/pull/230)
+- Show replies in invite screen [PR #232](https://github.com/marmot-protocol/whitenoise/pull/232)
 
 ### Changed
 
@@ -69,7 +70,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixes logout not working after app reinstall [PR #31](https://github.com/marmot-protocol/sloth/pull/31)
 - Fixes sign out exception and adds dedicated sign out screen with private key backup [PR #45](https://github.com/marmot-protocol/sloth/pull/45)
 - QR code color now uses theme-aware color [PR #183](https://github.com/marmot-protocol/sloth/pull/183)
-- DM avatar color inconsistency [PR #199](https://github.com/marmot-protocol/sloth/pull/199)
+- DM avatar color inconsistency [PR #199](https://github.com/marmot-protocol/sloth/pull/199), [PR #232](https://github.com/marmot-protocol/whitenoise/pull/232)
 - Ignore duplicate newMessage for accounts on same device [PR #244](https://github.com/marmot-protocol/whitenoise/pull/244)
 - Sanitize malformed UTF-16 in user metadata to prevent rendering crashes [PR #234](https://github.com/marmot-protocol/whitenoise/pull/234)
 - Lock app orientation to portrait mode [PR #235](https://github.com/marmot-protocol/whitenoise/pull/235)